Extract API response behavior to a controller mixin
