+++ controller/lib/controller_engine.rb
@@ -39,8 +39,8 @@ class OpenShift::Responder < ::ActionController::Responder
   ACTIONS_FOR_VERBS	=	{ :post => :new, :put => :edit, :patch => :update }
   def api_behavior(error)
     raise error unless resourceful?
-    status = resource.each{ |r| break(r[:status]) if r.class == Hash && r.has_key?(:status) }
-    display resource[0], status: status
+    status = options[:status] || (resource.respond_to?(:status) && resource.status) || nil
+    display resource, status: status
   end
 end
