+++ controller/lib/openshift/controller/api_behavior.rb
@@ -1,0 +1,8 @@
+module OpenShift
+  module Controller
+    module ApiBehavior
+      extend ActiveSupport::Concern
+    end
+  end
+end
+
